--- tests/Domain/Experience/Reactors/PullRequestExperienceReactorTest.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Tests\Domain\Experience\Reactors;
-
-use Tests\TestCase;
-
-class PullRequestExperienceReactorTest extends TestCase
-{
-}
